refactor(eigendb): renamed 'vector/' directory to reflect the endpoints defined inside
